Started db code
Made Find and Query classes for the new type of TFFI. This TFFI sends a request that instructs the server to do a query of the db so we can search for destinations

Updated find and query to match TFFI definition. Modified DB manager to use command line args to log in

Got db querying working with a search string. Build list of places for TFFI from that so we can send it back to the client. Almost done with #163 and #165

Finished coverage tests for Find.java. Future plan is to move Plan.java and Query.java into a separate package so we can do JUnit tests on them using a library that can test a Spark server with HTTP. Closes #163 and closes #166

CodeClimate

CodeClimate